+++ CHANGELOG.md
@@ -7,7 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
-[Unreleased]: https://github.com/atomist/slack-messages/compare/0.3.0...HEAD
+[Unreleased]: https://github.com/atomist/slack-messages/compare/0.3.1...HEAD
+
+## [0.3.1] - 2017-05-10
+
+[0.3.1]: https://github.com/atomist/slack-messages/compare/0.3.0...0.3.1
+
+### Fixed
+-   **BREAKING** `ButtonSpec` should be an interface
 
 ## [0.3.0] - 2017-05-10